+++ main/point_of_sale.py
@@ -1,0 +1,11 @@
+class Display:
+    def get_latest(self):
+        return ""
+
+
+class PointOfSaleSystem:
+    def __init__(self, display: Display):
+        pass
+
+    def on_barcode(self, barcode: str):
+        pass
