+++ tests/test_point_of_sale.py
@@ -1,14 +1,4 @@
-class Display:
-    def get_latest(self):
-        return ""
-
-
-class PointOfSaleSystem:
-    def __init__(self, display: Display):
-        pass
-
-    def on_barcode(self, barcode: str):
-        pass
+from main.point_of_sale import Display, PointOfSaleSystem
 
 
 def test_empty_barcode():
